server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0,<6",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-environ>=0.11",
    "psycopg[binary]>=3.1",
    "google-auth>=2.29",
    "django-storages>=1.14",
    "django-extensions>=4.1",
]

[project.optional-dependencies]
dev = [
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "ipython>=8.24",
]
production = [
    "gunicorn>=22.0",
]
